fix `dbt docs generate` for athena execution engines 1 & 2.
execution engine 2 is being rolled out in us-east-1 (and a few other regions) currently.
soon, it will become the default engine athena uses worldwide.

there are some sql-parsing incompatibilities between the two versions when tables are
joined like this

```
form table1
join table2 using (col1, col2)
```

in version 1, it was necessary to specify `col1` as either `table1.col1` or `table2.col1`
in the where clause or select clause. however in version 2, that generates an error

to ensure compatibility across both versions, the simplest thing to do is to avoid specifying
the join as `using (col1, col2)`

this pattern is currently used in fetching the database metadata when executing `dbt docs generate`, so
I was not able to build docs against athena running on execution engine version 2.

with the change, `dbt docs generate` works successfully across both engine versions.